You should catch exceptions while constructing a StepBody/ExecutionBody Fixes #348
